chore: move settings spec into main dir
